Stylistic changes, font style changes, theme changes, settings organization changes, and real time weather overview updates, Version 3.1

+++ src/utils/weatherCodes.js
@@ -270,5 +270,5 @@ export function formatDay(dateString) {
   tomorrow.setDate(today.getDate() + 1)
   if (date.toDateString() === today.toDateString()) return 'Today'
   if (date.toDateString() === tomorrow.toDateString()) return 'Tomorrow'
-  return date.toLocaleDateString('en-US', { weekday: 'long' })
+  return date.toLocaleDateString('en-US', { weekday: 'short' })
 }
